requirements.txt
selenium
python-dotenv
pydantic
flask
schedule
requests
etsyv3
